+++ src/index.js
@@ -1,4 +1,4 @@
-import Avatar, { AvatarName, AvatarStack } from './components/avatar';
+import Avatar, { AvatarInitials, AvatarStack } from './components/avatar';
 import Badge from './components/badge';
 import Banner from './components/banner';
 import Box from './components/box';
@@ -66,7 +66,7 @@ import {
 
 export {
   Avatar,
-  AvatarName,
+  AvatarInitials,
   AvatarStack,
   AsyncSelect,
   Badge,